Fixed the horrible bug, added strafing, set framerate
Well, that horrible bug is fixed. Turns out it was an issue with Eigen's vector alignment. The program seems to be stable now, I rammed my face into the walls over and over and nothing bad happened. The framerate is now set to 30 frames per second, rather than fluctuating inconsistently.

After a few small changes and looking over the code to ensure that Eigen will not produce future oddities, I will perhaps texture the walls to make the world a little more pleasant to look at.